==> HousePrice/0_Concated_Data.ipynb <==
import pandas as pd

# Danh sách các tệp CSV cần gộp
file_paths = ['D:/DAP391m/data/house30-50.csv', 'D:/DAP391m/data/house50-60.csv', 'D:/DAP391m/data/house60-80.csv']

# Khai báo một DataFrame trống để lưu trữ dữ liệu từ các tệp CSV
merged_data = pd.DataFrame()

# Đọc từng tệp CSV và thêm dữ liệu vào DataFrame gộp
for file_path in file_paths:
    data = pd.read_csv(file_path)
    merged_data = pd.concat([merged_data, data], ignore_index=True)

# Lưu DataFrame gộp thành một tệp CSV mới
merged_data.to_csv('merged_file.csv', index=False)

df = pd.read_csv('merged_file.csv')
df

import pandas as pd

# Danh sách các tệp CSV cần gộp
file_paths = ['D:/data/house.csv', 'D:/DAP391m/data/house1.csv', 'D:/DAP391m/data/house2.csv', 'D:/DAP391m/data/house3.csv','D:/DAP391m/data/house4.csv',
             'D:/DAP391m/data/house5.csv', 'D:/DAP391m/data/house6.csv']

# Khai báo một DataFrame trống để lưu trữ dữ liệu từ các tệp CSV
merged_data = pd.DataFrame()

# Đọc từng tệp CSV và thêm dữ liệu vào DataFrame gộp
for file_path in file_paths:
    data = pd.read_csv(file_path)
    merged_data = pd.concat([merged_data, data], ignore_index=True)

# Lưu DataFrame gộp thành một tệp CSV mới
merged_data.to_csv('van_file.csv', index=False)

df = pd.read_csv('van_file.csv')
df

file_paths = ['D:/DAP391m/merged_file.csv', 'D:/DAP391m/van_file.csv']

# Khai báo một DataFrame trống để lưu trữ dữ liệu từ các tệp CSV
merged_data = pd.DataFrame()

# Đọc từng tệp CSV và thêm dữ liệu vào DataFrame gộp
for file_path in file_paths:
    data = pd.read_csv(file_path)
    merged_data = pd.concat([merged_data, data], ignore_index=True)

# Lưu DataFrame gộp thành một tệp CSV mới
merged_data.to_csv('final_file.csv', index=False)

df = pd.read_csv('final_file.csv')
df

import pandas as pd
file_paths = ['D:/DAP391m/final_file.csv', 'D:/DAP391m/data/house_price_1.csv', 'D:/DAP391m/data/house_price_2.csv']

# Khai báo một DataFrame trống để lưu trữ dữ liệu từ các tệp CSV
merged_data = pd.DataFrame()

# Đọc từng tệp CSV và thêm dữ liệu vào DataFrame gộp
for file_path in file_paths:
    data = pd.read_csv(file_path)
    merged_data = pd.concat([merged_data, data], ignore_index=True)

# Lưu DataFrame gộp thành một tệp CSV mới
merged_data.to_csv('hanoi_house.csv', index=False)

df = pd.read_csv('hanoi_house.csv')
df

import pandas as pd
file_paths = ['D:/DAP391m/hanoi_house.csv', 'D:/DAP391m/data/house_price_3.csv', 'D:/DAP391m/data/house_price_4.csv']

# Khai báo một DataFrame trống để lưu trữ dữ liệu từ các tệp CSV
merged_data = pd.DataFrame()

# Đọc từng tệp CSV và thêm dữ liệu vào DataFrame gộp
for file_path in file_paths:
    data = pd.read_csv(file_path)
    merged_data = pd.concat([merged_data, data], ignore_index=True)

# Lưu DataFrame gộp thành một tệp CSV mới
merged_data.to_csv('hanoi_house_2.csv', index=False)

df = pd.read_csv('./hanoi_house_2.csv')
df
