==> HousePrice/02_Data_Clean.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('./hanoi_house_2.csv')

df.shape

df = df[~df['price'].str.contains('Giá thỏa thuận|triệu')]
df['price'] = df['price'].str.replace('[+tỷ]', '', regex=True)
df['price'] = df['price'].replace(',', '.', regex=True)
df['price'] = df['price'].astype(float)

df['duong_vao'] = df['duong_vao'].str.replace('[+m]', '', regex=True)
df['duong_vao'] = df['duong_vao'].replace(',', '.', regex=True)
df['duong_vao'] = df['duong_vao'].astype(float)
df['duong_vao'].fillna(value=df['duong_vao'].mean().round(1), inplace=True)

df['floor'] = df['floor'].str.replace('[+tầng]', '', regex=True)
df['floor'] = df['floor'].replace(',', '.', regex=True)
df['floor'] = df['floor'].astype(float)

df['size'] = df['size'].str.replace('[+m²]', '', regex=True)
df['size'] = df['size'].replace(',', '.', regex=True)
df['size'] = df['size'].astype(float)

df = df[~df['Price_per_m2'].str.contains('nghìn')]
df['Price_per_m2'] = df['Price_per_m2'].str.replace('[+tr/m²]', '', regex=True)
df['Price_per_m2'] = df['Price_per_m2'].replace(',', '.', regex=True)
df['Price_per_m2'] = df['Price_per_m2'].astype(float)


df['mat_tien'] = df['mat_tien'].str.replace('[+m]', '', regex=True)
df['mat_tien'] = df['mat_tien'].replace(',', '.', regex=True)
df['mat_tien'] = df['mat_tien'].astype(float)

df.drop(columns=['title', 'link_item', 'link_item.1', 'bed_idx', 'toi_idx', 'index_', 'Unnamed: 0'], inplace=True)

locations_to_exclude = ['Quận 7, Hồ Chí Minh', 'Bình Thạnh, Hồ Chí Minh', 'Bắc Giang, Bắc Giang', 'Thuận An, Bình Dương', 'Quận 2, Hồ Chí Minh', 'Quận 10, Hồ Chí Minh']
df = df[~df['location'].isin(locations_to_exclude)]


df.shape

df

df['phap_ly'].value_counts()

# hợp đồng
my_list = ['Hợp đồng mua bán', 'Hợp đồng mua bán.']

# Sổ đỏ - sổ hồng
my_list1 = ['SỔ ĐỎ NÉT','Sổ đỏ đầy đủ', 'Sổ đỏ đẹp chính chủ', 'Sổ đỏ vuông đẹp sẵn sàng', 'Pháp lý chuẩn sổ đỏ sẵn sàng giao dịch ngay',
           'Sổ đỏ, sổ hồng', 'Sổ đỏ vuông đẹp', 'Pháp lý chuẩn chỉ, sổ đỏ chính chủ riêng biệt, đất sạch không tranh chấp.',
           'Sổ hổng.', '- Sổ đỏ chính chủ, riêng biệt, pháp lý rõ ràng, không tranh chấp. Có đầy đủ giấy phép xây dựng, hồ sơ thiết kế, sơ đồ điện nước đầy đủ.',
           'Sổ đỏ, chính chủ.', 'Sổ đỏ pháp lý chuẩn sẵn sàng giao dịch', 'Đã Có sổ đỏ', 'Sổ sạch chính chủ, vuông đẹp, pháp lý chuẩn cất Két',
           'Sổ sạch 1 chủ, pháp lý chuẩn','ổ chính chủ giao dịch ngay','Sổ đỏ riêng chính chủ.', 'Sổ đỏ chính chủ riêng.',
           '* Pháp lý rõ ràng không tranh chấp, chứng thực đầy đủ, sổ đỏ chính chủ chuyển nhượng ngay. Đất sạch, mặt tiền rộng thoáng rộng đắc địa tài lộc, không lỗi phong thủy.',
           'Sổ đỏ cất két, pháp lý sạch', 'Giấy tờ pháp lý rõ ràng, sổ đỏ chính chủ','Sổ đỏ nét','Sổ chính chủ vuông đẹp','Đã có sổ đỏ pháp lý đầy đủ, không tranh chấp chính chủ.',
           'Chính chủ, pháp lý Chuẩn','Sổ đỏ cất két. Giao dịch trong ngày.','Sổ đỏ cất két sẵn sàng giao dịch.','Sổ đẹp', 'Sổ đỏ đẹp.','Sổ đỏ chính chủ chờ giao dịch .',
           'SỔ ĐỎ CHÍNH CHỦ', 'Sổ cất két', 'Sổ đỏ chính chủ nở hậu', 'Sổ hồng chính chủ', 'Sổ đỏ chính chủ riêng biệt, pháp lý sạch, không tranh chấp. Giấy phép xây dựng đầy đủ.', 
           'Pháp lý Chuẩn, sẵn sàng giao dịch', 'Sổ sạch, pháp lý chuẩn.', 'Sổ chính chủ vuông vắn.', 'Sổ sạch tên 2 vc cất Két chờ trao chủ mới',
           'Sổ chính chủ nở hậu', 'Sổ đỏ lâu dài.', 'Sổ đỏ chính chủ, sẵn sang tên.']

for i in my_list1:
    df['phap_ly'] = df['phap_ly'].replace(i, 'Sổ đỏ/ Sổ hồng')

# chưa sổ
my_list2 = ['Đang chờ sổ', 'Chờ.', 'Đang chờ sổ.', 'Chưa sổ.']

for i in my_list2:
    df['phap_ly'] = df['phap_ly'].replace(i, 'Chưa')

df['phap_ly'].value_counts()

df

df['noi_that'].value_counts()

df['noi_that'].value_counts().sum()

# Đầy đủ
my_list1 = ['FULL nội thất sang trọng','Đủ','FULL NỘI THẤT','Tặng hết nội thất sang xịn cho khách thiện chí.','Cửa 2 lớp, bên trong cửa kính cường lực bên ngoài cửa cuốn, cầu thang và cửa thông phòng gỗ Lim Nam Phi, thiết bị vệ sinh Inax, trần thạch cao các phòng, gạch chống xước, đầu chờ điều hòa,',
           'Đẹp, mới tinh, chỉ việc đến ở','Nội thất liền tườnh','Đầy đủ nội thất sang xịn.','Sang Trọng','Tặng toàn bộ nội thất gắn tường sang đẹp.',
           'Sẵn nội thất','Đầy đủ nội thất','Nội thất đầy đủ.','Đầy đủ để lại toàn bộ nội thất cho người mới. Riêng bộ sofa 45tr 3 đh daikin mới cứng. Giường tủ tivi sony.',
           'Nhà đã hoàn thiện đầy đủ nội thất, chỉ việc dọn về ở.','- Nội thất: Cửa chính, cửa cuốn và 1 cửa phụ, cửa thông phòng, bậc cầu thang làm bằng gỗ Lim. Cửa kính thủy lực ngoại nhập, tủ bếp, tủ âm tường đựng quần áo, điều hòa các phòng bình nóng lạnh, thiết b',
           '- Nội thất: Đầy đủ nội thất hoàn thiện, chỉ việc về ở. - Cửa và cầu thang gỗ Lim, đầy đủ tủ bếp, thiết bị vệ sinh, cửa chính 2 lớp...',
           'Đầy đủ Nội thất siêu víp gỗ óc chó, điều hoà âm Trần, thiết bị điện panasonic, thiết bị wc inax, bồn sục, cầu thang đá marber, sàn gỗ xương cá.',
           'Tặng full nội thất đẹp','Full cơ bản đầy đủ','Đã có đầy đủ các thiết bị cơ bản như trần thạch cao, đèn điện Led, dây điện Trần Phú, thiết bị vệ vệ sinh cao cấp Viglacera, Inax...',
           'Đầy đủ đầy đủ nội thất giường tủ, điều hòa.','Đầy đủ đẹp','Full nội thất đẹp.','Chủ tặng lại toàn bộ nội thất xịn.',
           'Tôi để lại toàn bộ nội thất xịn do chính 2 vợ chồng tôi sắm sửa cho người có duyên! Người mua chỉ việc đến ở, không phải sắm sửa thêm gì cả.',
           'Full nội thất đẹp: điều hòa, tủ bếp....','- Full nội thất mới.','- Nội thất để lại.','Nội thất đẩy đủ','Tặng toàn bộ nội thất','Đầy đủ nội thất gắn tường cơ bản.',
            'Nội thất bao gồm tủ bếp, đèn chùm, điều hòa, nóng lạnh, kệ tivi, tủ rượu, bếp từ, hút mùi cửa chính, cửa thông phòng gỗ Lim, toàn bộ hệ thống thiết bị vệ sinh, đèn chiếu sáng, trang trí đầy đủ,...',
            'Đầy Đủ','Tặng lại toàn bộ nội thất: Giường, bàn ghế, ti vi, tủ lạnh, điều hoà...','Đã hoàn thiện sẵn nội thất cơ bản bao gồm: Tủ bếp gỗ Sồi Nga, hệ thống vệ sinh Inax, cầu thang tay vịn gỗ, trần thạch cao, cửa thông phòng gỗ Lim, đèn trang trí và chiếu sáng, máy bơm nước, ch.',
            'Trang bị nội thất cơ bản: Tủ bếp gỗ tự nhiên, hệ thống cửa nhôm kính Việt Pháp cách âm cách nhiệt, cầu thang tay vịn gỗ, gạch chống xước, hệ thống đèn chiếu sáng trang trí hiện đại đẹp mắt,...',
            'Full nội thất sang trọng','Chủ để lại toàn bộ nội thất','Tặng nội thất','Mới đẹp đầy đủ','Nội thất cao cấp đầy đủ.'
           'Full nội thất xịn.', 'Nội thất liền tường','Full nội thất xịn.']

for i in my_list1:
    df['noi_that'] = df['noi_that'].replace(i, 'Đầy đủ')

# cơ bản
my_list2 = ['Cơ bản.', 'Cơ bản, hiện đại, bắt mắt','Nội thất có bản','Nội thất cơ bản, điều hòa, nóng lạnh.',
           'Có bản', 'Full cơ bản','Full đầy đủ nội thất cao cấp+ Thang Máy nhập khẩu.','Còn mới','Nội thất đẹp.',
           'Nội thất cơ bản.','Nội thất cơ bản','Có','Buồng tắm đứng đầy đủ thiết bị + WC khép kín.','Cơ bản đầy đủ.','']

for i in my_list2:
    df['noi_that'] = df['noi_that'].replace(i, 'Cơ bản')

# Cao cấp
my_list3 = ['Full nội thất cao cấp.','Toàn bộ nội thất xịn cao cấp','Đẹp Long Lanh, Nhập Khẩu','Cao cấp.','Nhập ngoại','Hiện đại',
           'Nội thất xịn sò.','Nội thất sang trọng','Nội thất xịn.','Nội thất nhập khẩu cao cấp.','Nội thất cao cấp']

for i in my_list3:
    df['noi_that'] = df['noi_that'].replace(i, 'Cao cấp')

# Không có
my_list4 = ['Không nội thất.','Không']

for i in my_list4:
    df['noi_that'] = df['noi_that'].replace(i, 'Không nội thất')

df

df.to_csv('hanoi_house_final_4.csv', index=False)
