==> HousePrice/05_Machine_Learning_Model.ipynb <==
import pandas as pd
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression, RidgeCV, LassoCV, ElasticNetCV
from sklearn.metrics import mean_squared_error, make_scorer, mean_absolute_error 
from sklearn.metrics import r2_score

df = pd.read_csv("./hanoi_house_processing_1.csv")

df

y=df['price']
X=df.drop(['price'], axis = 1)
X

## Spliting Data

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

### Linear Regression

reg = LinearRegression().fit(X_train, y_train)

y_train_pred = reg.predict(X_train)
y_test_pred = reg.predict(X_test)

r2_score(y_test, y_test_pred)

print("Absolute Error", mean_absolute_error(y_test,y_test_pred))
print("Squared Error", mean_squared_error(y_test, y_test_pred)**(0.5))

### Support Vector Regression

from sklearn.svm import SVR
from sklearn.model_selection import GridSearchCV

svr_model = SVR(kernel='linear')
svr_model.fit(X_train, y_train)

y_train_svm_pred = svr_model.predict(X_train)
y_test_svm_pred = svr_model.predict(X_test)

r2_score(y_test, y_test_svm_pred)

print("Absolute Error", mean_absolute_error(y_test,y_test_svm_pred))
print("Squared Error", mean_squared_error(y_test, y_test_svm_pred)**(0.5))

### Random Forest

from sklearn.model_selection import RandomizedSearchCV
from sklearn.ensemble import RandomForestRegressor
import numpy as np

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 200, stop = 2000, num = 10)]
# Number of features to consider at every split
max_features = ['auto', 'sqrt', 'log2']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(10, 110, num = 11)]
max_depth.append(None)
# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]
# Method of selecting samples for training each tree
bootstrap = [True, False]
# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}
            
     

# Use the random grid to search for best hyperparameters
# First create the base model to tune
rf = RandomForestRegressor()
# Random search of parameters, using 3 fold cross validation, 
# search across 100 different combinations, and use all available cores
rf_random = RandomizedSearchCV(estimator = rf, param_distributions = random_grid, n_iter = 10, cv = 3, verbose=2, random_state=42, n_jobs = -1)


# Fit the random search model
rf_random.fit(X_train, y_train)

rf_random.best_params_

model_final = rf_random.best_estimator_

model_final.get_params()

y_train_rf_pred = model_final.predict(X_train)
y_test_rf_pred = model_final.predict(X_test)

r2_score(y_test, y_test_rf_pred)

print("Absolute Error", mean_absolute_error(y_test,y_test_rf_pred))
print("Squared Error", mean_squared_error(y_test, y_test_rf_pred)**(0.5))

import pickle
pickle.dump(model_final,open('./model/model_rf.pickle',mode='wb'))

pickle.dump(svr_model,open('./model/svr_model.pickle',mode='wb'))
pickle.dump(reg,open('./model/model_Linear.pickle',mode='wb'))

### Decision Tree

from sklearn.tree import DecisionTreeRegressor 
from sklearn.model_selection import GridSearchCV

param_grid = {'max_depth': [5, 10, None], 'max_features': ['auto', 'log2', 'sqrt'], 'min_samples_leaf': [5, 6, 7, 9,11, 12, 13, 15]}
reg_decision_model = DecisionTreeRegressor(random_state = 42) 

reg_decision_model.get_params().keys()

regressor = GridSearchCV(reg_decision_model,param_grid=param_grid,scoring='neg_mean_absolute_error',cv=3,verbose=3)
regressor.fit(X_train, y_train)

regressor.best_params_

model_dt_final = regressor.best_estimator_


y_train_dt_pred = model_dt_final.predict(X_train)
y_test_dt_pred = model_dt_final.predict(X_test)

r2_score(y_test, y_test_dt_pred)

print("Absolute Error", mean_absolute_error(y_test,y_test_dt_pred))
print("Squared Error", mean_squared_error(y_test, y_test_dt_pred)**(0.5))

pickle.dump(model_dt_final,open('./model/model_dt.pickle',mode='wb'))

-----
**END**
