==> HousePrice/06_Predict.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('./house_test.csv')

df.shape

df = df[~df['price'].str.contains('Giá thỏa thuận|triệu')]
df['price'] = df['price'].str.replace('[+tỷ]', '', regex=True)
df['price'] = df['price'].replace(',', '.', regex=True)
df['price'] = df['price'].astype(float)

df['duong_vao'] = df['duong_vao'].str.replace('[+m]', '', regex=True)
df['duong_vao'] = df['duong_vao'].replace(',', '.', regex=True)
df['duong_vao'] = df['duong_vao'].astype(float)
df['duong_vao'].fillna(value=df['duong_vao'].mean().round(1), inplace=True)

df['floor'] = df['floor'].str.replace('[+tầng]', '', regex=True)
df['floor'] = df['floor'].replace(',', '.', regex=True)
df['floor'] = df['floor'].astype(float)

df['size'] = df['size'].str.replace('[+m²]', '', regex=True)
df['size'] = df['size'].replace(',', '.', regex=True)
df['size'] = df['size'].astype(float)

df = df[~df['Price_per_m2'].str.contains('nghìn')]
df['Price_per_m2'] = df['Price_per_m2'].str.replace('[+tr/m²]', '', regex=True)
df['Price_per_m2'] = df['Price_per_m2'].replace(',', '.', regex=True)
df['Price_per_m2'] = df['Price_per_m2'].astype(float)


df['mat_tien'] = df['mat_tien'].str.replace('[+m]', '', regex=True)
df['mat_tien'] = df['mat_tien'].replace(',', '.', regex=True)
df['mat_tien'] = df['mat_tien'].astype(float)

df.drop(columns=['title', 'link_item', 'link_item.1', 'bed_idx', 'toi_idx', 'index_', 'Unnamed: 0'], inplace=True)

df['location'].value_counts()

df

df['phap_ly'].value_counts()

my_list2 = ['Sổ đỏ/ Sổ hồng.', 'Sổ đỏ', 'Sổ đỏ chính chủ', 'Sổ đỏ.', 'Sổ đỏ chính chủ.', 'Sổ đỏ chính chủ, mới tinh.','Chuẩn','Đã có sổ']
for i in my_list2:
    df['phap_ly'] = df['phap_ly'].replace(i, 'Sổ đỏ/ Sổ hồng')

df['noi_that'].value_counts()

my_list2 = ['Đầy đủ.','Mới','Đầy đủ nội thất hiện đại, thang máy nhập khẩu xịn, cầu thang gỗ giữa, thang máy giữa, sàn ép gỗ tốt.','Cơ bản đầy đủ.']
for i in my_list2:
    df['noi_that'] = df['noi_that'].replace(i, 'Đầy đủ')

my_list3 = ['Cơ bản.']
for i in my_list3:
    df['noi_that'] = df['noi_that'].replace(i, 'Cơ bản')

my_list4 = ['Nội thất cao cấp nhập khẩu hoàn toàn. Cửa phòng ngủ, cầu thang, gỗ Lim nguyên khối, thiết bị vệ sinh Inax, tủ bếp gỗ cao cấp, trần thạch cao.']
for i in my_list4:
    df['noi_that'] = df['noi_that'].replace(i, 'Cao cấp')

df['noi_that'].value_counts()

df

df.to_csv('hanoi_house_test.csv', index=False)

df = pd.read_csv('./hanoi_house_test.csv')

df

def replace_missings (X) :
    X=X.drop(X.isna().sum()[X.columns[X.isna().sum() > 1000]].index, axis = 1)
    lis = X.isna().sum()[X.columns[X.isna().sum() > 1]].index
    for k in lis : 
        if type(X[k].iloc[0]) == str:
            X[k]=  X[k].fillna(X[k].iloc[0], inplace=True)
        
        else :
            X[k]=  X[k].fillna(X[k].mean())
    return X 

def labelencoding (X) :
    lisle = []
    for col in X.columns:
        if X[col].dtype == 'object':
            label_encoder = LabelEncoder()
            X[col] = label_encoder.fit_transform(X[col])
            lisle.append(label_encoder)
    return X

def remove_outliers(X, threshold):
    z_scores = (X - X.mean()) / X.std()
    outliers = (z_scores.abs() > threshold).any(axis=1)
    X = X[~outliers]
    return X

datatrain = pd.read_csv('./hanoi_house_test.csv')
datatrain= labelencoding(datatrain)
datatrain
datatrain= replace_missings (datatrain)
datatrain = remove_outliers(datatrain, 5)
datatrain

datatrain.to_csv('./hanoi_house_test_1.csv')

y_test=datatrain['price']
X_test=datatrain.drop(['price'], axis = 1)
X_test

# Load all models
import pickle
model_rf =  pickle.load(open('./model/model_rf.pickle',mode='rb')) # machine learning model (SVM)

y_test

y_predict = model_rf.predict(X_test)

y_predict

final = pd.DataFrame({'y_test': y_test, 'y_predict': y_predict})

final


final.to_csv('./prediction_result.csv')
